pyproject.toml
[project]
name = "asyncio-implementation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "grpcio>=1.76.0",
    "sqlalchemy>=2.0.45",
]
